Split class into multiple files
